Update keyboard.js
Case statements make parsing data like this a lot easier in the long run. Neater to read and understand and not nearly as many if statements. A lot shorter too. You don't have to scroll through a bunch of junk now. You will have to change the main file as well but that's just a simple search and replace.